backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
langchain==0.1.0
langchain-openai==0.0.8
langchain-community==0.0.13
openai>=1.10.0,<2.0.0
python-dotenv==1.0.0
chromadb==0.4.22
tiktoken==0.5.2